requirements.txt
fusepy
qrcode
pyotp
bcrypt
pycrypto
simple-crypt
